Fix some signed/unsigned warnings

+++ src/shader/buffer/UniformBuffer.cpp
@@ -13,13 +13,13 @@ UniformBuffer::~UniformBuffer() {
 }
 
 GLint UniformBuffer::allocateBindPoint() {
-    if (allocatedBindPoints.size() == MaxBuffers()) {
+    if (allocatedBindPoints.size() == (size_t)MaxBuffers()) {
         // TODO: Better exception implementation.
         // char buff[1024];
         // snprintf(buff, sizeof(buff), "Uniform Buffer Bindings exceeded!\nLimit = %d\n\nsdl_exp UniformBuffer objs are not designed for sharing buffer bindings.", MaxBuffers());
         throw std::exception();
     }
-    for (unsigned int i = MaxBuffers() - 1; i >= 0; --i) {
+    for (unsigned int i = (size_t)MaxBuffers() - 1; i >= 0; --i) {
         if (allocatedBindPoints.find(i) == allocatedBindPoints.end()) {
             allocatedBindPoints.insert(i);
             return i;
